fix: HomeViewModel — defensiva en CurrentRoutineId y recarga tras sesión completada
- Defensive read de CurrentRoutineId (evita ClassCastException por tipo int
  guardado en versiones anteriores)
- CurrentRoutineId guardado como long en todos los puntos de escritura
- Null-check explícito de Routine.CurrentDay antes de llamar a getRoutineDayAsync
- _isInitialized movido al final de InitializeAsync (solo marca éxito real)
- AppearingAsync resetea _isInitialized si CurrentRoutineId no está en Preferences,
  forzando recarga del día y progreso semanal al volver de una sesión completada

+++ IFit/Services/TrainingService.cs
@@ -41,7 +41,7 @@ namespace IFit.Services
                 Debug.WriteLine($"Error creando rutina para el usuario {userId}: {response.ErrorMessage}");
                 return null;
             }
-            Preferences.Set("CurrentRoutineId", response.Data?.Id ?? 0L); // Guardar el ID de la rutina creada para futuras referencias
+            Preferences.Set("CurrentRoutineId", (long)(response.Data?.Id ?? 0)); // Guardar el ID de la rutina creada para futuras referencias
 
             return response.Data;
         }
